Changed all ints to uints where appropriate.  Added image bounds checking in ImageReader.read().

+++ PyramidLib/AsynchronousPyramidGenerator.cs
@@ -10,23 +10,23 @@ namespace Pyramid.Generator
 {
 	public class AsynchronousPyramidGenerator : PyramidGenerator
 	{
-        private int levels;
+        private uint levels;
         private TileReader tileReader;
 		
-        public AsynchronousPyramidGenerator(string filename, int levels)
+        public AsynchronousPyramidGenerator(string filename, uint levels)
 		{
             this.levels = levels;
 
-            int tilesPerSide = (int)Math.Pow(2, levels - 1);
+            uint tilesPerSide = (uint)Math.Pow(2, levels - 1);
 
             ImageReader imageReader = new ImageReader(filename);
 
             // for now get the smaller of the image width and height; this will mean that
             // some of the longer dimension will be excluded from the final pyramid tiles
-            int length = Math.Min(imageReader.Height, imageReader.Width);
+            uint length = (uint)Math.Min(imageReader.Height, imageReader.Width);
 
             // calculate the tile dimension
-            int tileDimension = length / tilesPerSide;
+            uint tileDimension = length / tilesPerSide;
 
             // finally, create the TileReader
             tileReader = new TileReader(imageReader, tileDimension);
@@ -45,7 +45,7 @@ namespace Pyramid.Generator
             createTileAsync(levels - 1, 0, 0);
         }
 		
-        private Bitmap createTileAsync(int level, int tileX, int tileY)
+        private Bitmap createTileAsync(uint level, uint tileX, uint tileY)
         {
             //Console.WriteLine("Creating tile. x: {0}, y: {1}, level: {2}", tileX, tileY, level);
 
@@ -60,8 +60,8 @@ namespace Pyramid.Generator
 			
             // create tasks for the four child tiles in the level below this one
             Bitmap scaledImage = null;
-            int[] xs = {2 * tileX, 2 * tileX + 1};
-            int[] ys = {2 * tileY, 2 * tileY + 1};
+            uint[] xs = {2 * tileX, 2 * tileX + 1};
+            uint[] ys = {2 * tileY, 2 * tileY + 1};
             using (var topLeft = Task<Bitmap>.Factory.StartNew(() => { return createTileAsync(level - 1, xs[0], ys[0]); }))
             using (var topRight = Task<Bitmap>.Factory.StartNew(() => { return createTileAsync(level - 1, xs[1], ys[0]); }))
             using (var bottomLeft = Task<Bitmap>.Factory.StartNew(() => { return createTileAsync(level - 1, xs[0], ys[1]); }))
